requirements.txt
streamlit
langchain
langchain-openai
beautifulsoup4
python-dotenv
chromadb
openai
python-dotenv
langchain-community
langchain
langchain-core
streamlit
langchain-cli